pyproject.toml
[project]
authors = [
  {name = "Christian Camacho", email = "chris.ca.dev@gmail.com"},
]
dependencies = [
  "fastapi[standard]>=0.129.2",
  "jinja2>=3.1.6",
  "powerwalk>=1.0.0",
  "jinja-markdown2>=0.4.0",
  "pyyaml>=6.0",
  "markdown>=3.0",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.20",
  "pyjwt>=2.11.0",
  "uvicorn[standard]>=0.41.0",
  "python-dotenv>=1.2.1",
  "htmlmin>=0.1.12",
]
description = "The idea is to connect website of logs with a live chat in a single page. This to generate interaction between people and content."
license = {text = "MIT"}
name = "blog-chat"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]

[tool.pdm.scripts]
prod = "uvicorn blog_chat.app:app --host 0.0.0.0 --port 9091"
start = "fastapi dev src/blog_chat/app.py"

[tool.pdm]
distribution = true
